move objects into own package
